requirements.txt
torch
torchvision
transformers
pillow
faiss-gpu
numpy
tqdm
